renderDate etc.: Added support for ical.js-ish Time instances.
This solution won't work for formatting times that don't exist in the browser's time zone, eg. 2016-3-27 2:30:00 in Europe/Copenhagen.

The reason is that an intermediate Date instance with the same field values is created, and such times cannot be represented. I have an idea for a fix, but it will take some time to get it right.